Project Structure: support alphanumeric SDK versions
Adds supoort for Project Structure to handle SDK versions that are
alphanumeric instead of merely numeric. This is done through a
INTEGER_OR_STRING BuildFileKeyType that is always represented
internally as string but marshals to either string or integer
literal in the Groovy file as appropriate. The UI deals with this
new type without modification.

Change-Id: I0b8170496dbb7fa99e4a262d73f2dedb5350910b