Fix Gradle-project based unit tests
First, this updates the unit test runner which loads and builds the
Gradle unit test projects such that it searches through the build.gradle
files in the project and replaces older dependencies (of the Gradle
plugin, and of the buildToolsVersion) with the minimum requirements.
This means that we don't have to go and update all the unit test
projects every time we rev the requirements.

It will also insert a custom Maven URL if set via the android.mavenRepoUrl
system property, which lets you run unit tests against a dev version
of the Android plugin.

Finally I did update the various metadata files manually since we've
gotten rid of the packageName methods; this avoids build warnings
about APIs going away. While I was at it I bumped the versions too;
this isn't necessary given that the project loader above will replace
them on the fly, but since I've changed the schema I thought it best
to make the projects directly loadable in Studio as well.

Change-Id: Ib88437ed8a7ade540d3f5b6705984f423c63d924