Added workaround for null ModuleExtendedModel.
In Studio 0.5.9 we removed the exclusion of top-level build
folder because, starting with Android Gradle plug-in
0.10.0, that folder is where common exploded aars are
stored, and if excluded, they will not be visible to the
IDE. If this happens, finding references in editors will
not work.

We obtain the path of the build folder from
ModuleExtendedModel, a Gradle model that IDEA obtains
when syncing a project. We worked under the assumption that
this model will never be null. This assumption is
proved wrong by bug
https://code.google.com/p/android/issues/detail?id=70490

This bug also affects Java library modules. We use
ModuleExtendedModel to obtain the output paths for both
sources and resources for the "main" and "test" source sets.

The workaround implemented in this CL works as follows, in
the case of a null ModuleExtendedModel:
- To get the path of the top-level build folder, we loop
  through all the excluded folders of a module, until we
  find one with name "build."
- To get the paths of the output folders, we do the same
  as described above, and use the path of the build folder
  to hard code the paths of the source and resource
  folders.
This workaround works as long as users do not customize
the name and path of the build folder, which is something
not very common.

Change-Id: Ia176aecc29571d48366c078d2625f4f3c6a2d419