Added support for parsing structured messages coming from Android Gradle plug-in.
In this CL, we invoke Gradle passing the JVM arg
"android.invoked.from.ide". The Android Gradle plug-in,
when detecting this argument, will format any
warning/error messages using the pattern:

message_type|gradle_project_path|message

As part of this CL, we parse this output, create
the appropriate message for the "Messages" window
and link to the build.gradle file in the module
with a matching Gradle path.

This CL also fixes broken tests in BuildOutputParserTest.

Change-Id: I2912467ef5dfcb2f5b7923d3c5290feda5613595