Include manifest package in refactoring when applicationId is specified
Until now, we've deliberately excluded the package= attribute in Manifest
files from being included in package rename refactoring, since you can
move your code, but changing the package attribute will change the
identity of your app (so app updates won't work etc).

However, if the user has specified an application id in the Gradle
build files (either in the default config or in one of the flavors),
then the manifest package is *only* used for the implementation
(to determine the package of the R and BuildConfig classes, and to
resolve relative class references in the manifest). When that's the
case, the package attribute should participate normally in refactoring
operations.

See
59376: Refactoring a class doesn't update AndroidManifest.xml
(and literally dozens of similar issues filed over the years)

Change-Id: Ib40105b1ec0f48a1c6e0515d7f05197e151f80d6