Removed ProjectSyncMessageDataService.
The only function of this service was collecting all
unresolved dependencies found during project sync and
showing them in the "Messages" window. The problem
is how overcomplicated this is done:

1. During project sync, AndroidGradleProjectResolver
   obtains all unresolved dependencies from the Android
   and Java models and stores them in DataNode objects.
2. IDEA's external system framework then will get those
   DataNodes and invoke the soon-to-be-removed service
   to display them in the "Messages" window.

All of this is completely unnecessary because the models
have the unresolved dependencies, and we can just show
those during project configuration.

By removing this service, storing the current session's
state and reading it in the next session will be a lot
easier to implement.

Change-Id: Ic7bd6a276af506cd5954024605565c548525ced3