Fixed bug where unresolved dependencies were not treated as errors.
The side effect of this bug was:
- Source generation was being triggered, even though it was
  destined to fail, due to unresolved dependencies.
- Editors did not show "Sync failed" editor notification.

The main issue here is when we add error messages for unresolved
dependencies, they are added asynchronously, and when
PostProjectSyncTasksExecutor checks for the presence of errors,
the error messages are not added yet, and it thinks that the
project sync was successful. Same thing happens to the editor
notifications.

This CL simply adds a property to the project that indicates
whether it has unresolved dependencies.

Change-Id: If4e4f020b66d79fb9f6a33e85dbfae98d32491a9