Fix Step Seeking on Path Entry
There's a bug when moving between paths that if the first step of
a path is invisible it will be shown when the path is started or
if the last step of a path is invisible, similarly it will be shown
when moving backwards onto that path.

This change fixes that bug by leveraging the existing mechanism for
seeking the proper step to navigate to.

Change-Id: I6fe5ba66e44b10574cc86474868fc688d7e69a23