Now Gradle builds are considered successful if no BuildExceptions were thrown.
Before we use the number of *parsed* errors to determine
if a build was successful or not. The problem with that
is if there is a buggy output parser that accidentally
marks a message as "error," the IDE will think that the
build failed. If running an app, the IDE will refuse
to run the app because it mistakenly things something is
wrong.

Change-Id: Ica0a9d624fcf77699c8cce2ad1eb997e9200d4a1