Added yet another way to guess location of top-level build folder.
In this CL the IDE tries to guess the location of the top-level
build folder when the user has changed its default location.

This workaround works as follows: by default the IdeaModule model
excludes 2 folders: the .gradle one and the build one. We checked
that the model has exactly 2 excluded folders, and considers the
one that is no .gradle the build folder.

Workaround for issue
https://code.google.com/p/android/issues/detail?id=70490

Please note that this workarounds will be removed once Gradle
provides the location of the build folder. We already requested
that.

Change-Id: I095856a0e7ae3dc2653d91c91efc743016482068