Misc fixes for preview platforms
First, there is a new method on AndroidVersion: getFeatureLevel(),
which for a normal platform returns getApiLevel(), but for a preview
platform returns getApiLevel() + 1. This helps clean up code which
deals with preview platforms since clients no longer need to worry
about preview platforms specifically. This method is used in a number
of places, such that for example configuration folder matching will
match -v folders with the expected final number of the preview
platform rather than its current API number.

Second, when compiling with a preview platform the Gradle model will
coerce the minSdkVersion and targetSdkVersion values to the same API
number. This is useful for run and launch code, where we deliberately
want to restrict the APP to execution environments with the same
preview.  However, this has the side effect that it also causes the
rest of the tools machinery -- lint API check, template version
requirements, and so on -- to also limit itself to the preview
platform level, which is not optimal. Therefore, this CL repurposes
the AndroidModuleInfo#getMinSdkVersion method to get the "default"
value of minSdkVersion (by looking at the defaultConfig and the
productFlavors in the current variant) and using that one instead of
the final merged result. There is a new getRuntimeMinSdkVersion()
method which is used for the launch-related code.

Change-Id: I74fa0f70803d4339b832d37e3eb779682251b3c1