Fixes empty build.gradle file when re-creating a deleted project.
To reproduce the probem do the following:
. Create an empty activity project.
. Close android studio and delete the project from disk
. Reopen studio and create a new project in the same location

Android Studio will then hit the cached version of the file system and things
go wrong. The main thing you'll see is an empty top-level build.gradle file.
This will fail building and the error seen on screen says:

<project_dir>/app/build.gradle
   Error:(1, 0) Plugin with id 'android' not found.

Of course clearing the cache and restaring Studio fixes the issue.

Change-Id: I8f012a074844f3fb1daeeccd1f481fa01a38e6ea