Integrates basic allocation tracking into Android Studio's DDMS panel.
When system property ALLOC_TRACK is set, adds ability to start/stop allocation
tracking and displays results in a sortable table with a stack trace pane below
in which the user can click on a project file to navigate to it in the code.
Shows disabled Group by allocation class checkbox, Filter text field, and
Include trace checkbox as placeholders for not-yet-implemented functionality.

Change-Id: I5bb51304b64cc831b180bbb2f71c14875ab5b068
Signed-off-by: Rebecca Chen <rechen@google.com>