launcher: do not allow running apks that are not signed
Android does not allow installing apks that are not signed. The
package manager will throw an error if when attempting to install
an unsigned package. We already catch this error and display an
error dialog to the user.

This CL moves this check earlier: Even before attempting to install
an apk, we can check whether it is not signed, and if so, throw
an error.

Change-Id: I5386e09efc12b571b7a14ebea0275954dba0e069