device chooser: offload work from EDT
The Device Chooser dialog was slow to respond many times. This turned
out to be because of the repeating alarm (every 500ms) that did a bunch
of work in the EDT thread. This CL initializes the alarm so that the
handler is invoked in a pooled thread and not the EDT.

Change-Id: I051627206d511a4d3aa1e173a3a656f5dbc2efe8