Clean up layout device display menu
The device display menu has gotten longer and longer with each Nexus device we've
added, and with the new Android Wear and Android TV devices the menu actually scrolls!
This CL moves the generic devices into a new submenu, places the Android TV and
Android Wear devices in separate menus, and cleans up the code a bit (the code used
to have to deal with device definitions coming from the (separately distributed) tools,
so it couldn't make assumptions about what devices are available; that's no longer the
case.

Change-Id: I814e0e06f124c9ed5bd807ddf6649061be989c38