Handle rewriting unsupported class versions from .jars as well
The layout library can already rewrite custom view code compiled with
a higher bytecode version when those classes are in an .aar file or
in a local output folder. However, it is also possible to run into
this scenario when the custom view is in a plain .jar file (e.g. if
you compile the support library and try to access DrawLayout). This
CL uses the same class rewriter in that scenario.

Change-Id: I3c41ffa2f8ddfbb55801c6ebd57aa845acc089ff