Project Structure dialog now loads libraries from SDK repositories instead of hard-code them.
The "Choose Library Dependency" dialog in the Project Structure dialog
loads its list of "common libraries" by reading the contents of the
SDK Maven Repositories, instead of hard-coding the list.

If for some reason the contents of those repositories cannot be found,
the dialog falls back to the hard-coded list.

Change-Id: Idbc8fff008a79cc7458e9c28e5a42db5230a016c