Fixed bug where SDK created for android-L pointed to android-20 folder.
The problem is that Studio had some "smart code" that instead of
recreating the SDK with the new path, it assign the path to an
existing SDK. The matching was done by API level, which is wrong.

The fix is to remove that functionality. We don't gain too much
by reusing SDKs, most of the time taken is during indexing of the
SDK. By just changing paths of the SDK does not prevent indexing.

Fixes:
https://code.google.com/p/android/issues/detail?id=72589

Change-Id: I884d0cb3ddeec6a21a370dc9ecccdb7426695484