Add quickfixes for various lint checks
This CL adds quickfixes for various lint checks. There is one new
quickfix implementation, which takes a regular expression and a new
value, and searches the lint error location for the given regular
expression and if found, replaces the regular expression group with
the given value. This quickfix implementation is used to implement 25
lint checks, since a lot of lint checks are very simple (e.g. for "you
should use 5sp instead of 5dp for this textSize" the lint check just
has to replace sp with dp). This CL also replaces some existing lint
quickfixes, such as the deprecated plugin name quickfix (which had its
own custom implementation) with a usage of the new regexp lint fix.

There are a couple of other notable special cases:

- As part of our migration to remove dynamic gradle versions by default,
  the quickfix for dynamic gradle version will look up the currently
  implied verison from a dynamic version and insert that

- The deprecated lint fix had been disabled a long time ago with a comment
  that IntelliJ already handles deprecation warnings, but these do different
  things (IntelliJ flags deprecated Java code, the lint one flags deprecated
  XML tags and attributes) so I restored it.

Change-Id: Ie5b35c3c99d839224716a5296956fadf7622abe2