Now we delete project state cache on 'project close,' if the last project sync failed.
If a project sync failed and if none of the build.gradle files changed,
the next time the project is open, Studio will restore the project state
from the disk cache. The problem is that there is no indication that
the last sync failed, and there is no error message(s). By removing the
cache, we force a sync on the next 'project open' event. An example of
this scenario is when a user deletes a SDK from the SDK Manager, and that
SDK is being used in an Android Studio project.

We could have stored any error messages with the cache, but error messages
can also have quick fixes, and restoring this state can be messy.

This CL also includes some code cleanups: removal of warnings,
added missing @NotNull annotations and some other minor cosmetic changes.

Change-Id: Ia6e167a5949242fae15c3ce91454933b6bc1defd