Added "Open 'Messages' View" to editor notification when sync fails.
Replaced "Open Error Log" link with "Open 'Messages' View" link.
The "Open Error Log" link was useful when we displayed sync errors
in notification balloons, since the balloons dissappeared from the
screen but the error message was also recorded in the Log view.

Now that we display sync errors in the "Messages" view, it does not
make sense to open the Log view. Instead, we need to open the
"Messages" view, so users can see what went wrong.

Change-Id: I23da8b8c092ddfff814a893776fdfc7d0ccc38bf