Fixed bug where notification hyperlinks do not show up in balloon notifications.
The problem is that to include hyperlinks as part of the error message,
we have to manually concatenate the original error message and the 'toString'
value of each hyperlink. This CL add methods to AndroidGradleNotification
that take care of this, and cleans up client code to use these new methods.

Change-Id: Ie71881b880c00300339a3822a807db062979a20f