Allows Project Structure to see signing configs on Windows
Fixes a bug with Project Structure (Windows-only) where it wouldn't
show signing configs in the flavors panel. It was depending on an
ApplicationManager.invokeLater call to populate the popup with
configs, but because the ModalityState was not properly set on the
call, it wasn't getting invoked until the (modal) dialog closed.
Oddly, this worked okay on Mac.

Bug: 72089
Change-Id: I32da81de7931b24f553db15770fc3dceae64749e