First step towards the memory profiling tool.
This first version adds a tool window with a timeline view of the
current levels of the java heap. It uses the debugger protocol's
HPIF chunk to monitor the values. In this implementation the
application being monitored is the first one found that matches one
of the application packages in the project.

The option to launch it will be enabled only if the experimental
feature flag is set and will show up under the Tools menu.

Change-Id: I36399f9e6278482e433304b00027b4cbaa52adf6