Show friendly labels for Project Structure API choices
In the project structure, for combo boxes that let the user choose an
API level, show friendly lables like we do in the New Project Wizard.
To do this, there's a new infrastructure that allows any combo box
to show one thing in the UI but write a different thing to the build
file; there are static mappings to maintain these. This is also
applied to the Java language level to make that a little bit
friendlier.

Bug: 70719

Change-Id: Icf9ecc33a8bfd3561e6b9941d10b4210893e471c