De-couples strings editor from individual strings.xml files.
Create a wrapper class LocalResourceRepositoryAsVirtualFile that allows
a local resource repository to be opened as its own file.
StringResourceEditor pulls out and shows only the string resources from
the local resource repository file it is passed.

A notification bar appears across the top of strings.xml files with a
link to the editor and an option to hide the notification for the
duration of the current session. The right-click menu for strings.xml
files in the project navigation panel and, in Android View, for the node
under which they are grouped also offers an option to open the editor.

Change-Id: Iae8da096ce21e093bc622128d9cb285b6fdc271d
Signed-off-by: Rebecca Chen <rechen@google.com>