Now IDE recognizes 'missing target' errors coming from older versions of the Gradle plug-in.
This CL fixes a bug where importing a project that uses Android
Gradle plug-in 0.9.+ and a platform that is not installed in the
SDK does not result in the "Install Missing Platform" hyperlink
in the 'Messages' window.

The issue was that the error parser was not recognizing the
error message coming from an older plug-in version.

Change-Id: I288f5b4ce589a382f09353db7a39c932cc992044