Project Structure: only modify values that were touched
In the Project Structure dialog, do finer-grained tracking of what
the user has modified -- track it on a per-key level instead of
per-screen. Then only write out the individual keys that were
modified instead of replacing entire swaths of the build file. This
will make it less likely to unexpectedly stomp on custom user script
in the build file when the user didn't modify a certain value and
wasn't expecting it to be modified.

Bug: 74117
Change-Id: Ib7a83bac04c06746248ef2e7e38aec58971e9523