Replace treeStructureProvider with nodeDecorator
We introduced a treeStructureProvider earlier (1c68971dea0) in order
to change the displayed name for the build folder in cases where it
is outside the project root. This provider was then enhanced
(in 49af431c) to customize the icons displayed for modules.

Both these use cases can be handled using node decorator instead of
a full tree provider. This CL removes that tree structure provider
and replaces it with two node decorators: one that updates the build
folder name and another for the module icons.

Change-Id: I8f41e699cddd0384c93f17e10ad9dfbcc7fd03f9
(cherry picked from commit 6ec816c056bb6aeedac6139903202c2f312c6a1d)