Makes string resources editable from strings editor.
All code that touches the string resource data now lives in
StringResourceDataController. The editor can call the updateData() method
of the controller to retrieve the most current data from the local resource
repository, and on successful data retrieval,
StringResourceEditor.onDataUpdated() is called so that the editor can
refresh the table view, etc.  When the user makes a change within the
editor, StringResourceDataController.doEdit(...) changes the data in the
appropriate file(s) and calls updateData().

Known issues: We do not check that the key the user enters is unique, and
we use the deprecated method AndroidFacet.getPrimaryResourceDir().

(1) How the editing works:
  (a) When the translatable attribute of a key is changed, updates the
      attribute in every locale for which the key has a value. Setting
      translatable to true causes the attribute to be deleted.
  (b) When the name of a key is changed, updates the name in every
      locale for which the key has a value. Setting the name to ""
      (after trimming) causes the key to be deleted. (We may want to
      warn the user before this happens.)
  (c) When a non-empty default value or translation is edited to a non-
      empty value, updates the text of the resource item accordingly.
  (d) When a default value or translation goes from empty to non-empty,
      creates a resource item with the appropriate values for the text
      and the translatable attribute.
  (e) When a default value or translation goes from non-empty to empty,
      deletes the corresponding resource item.
  (f) Creates values directories and string resource files as necessary,
      although the editor does not currently use this functionality.

(2) Change to the cell detail panel (renamed to the edit panel): The Key
and Default Value text components are editable when a table cell is
selected, and the Translation text component is editable when a
translation table cell is selected.

(3) Misc: Uses AndroidResourceUtil.getDefaultResourceFileName(ResourceType)
to slightly simplify the way we check for a string resource file.

Change-Id: I6c8c1ccf01e9f09173125a8fcedae76588cbbbd9
Signed-off-by: Rebecca Chen <rechen@google.com>