Fixed issue where resources in AAR dependencies were not found.
Apparently Android Gradle plug-in v0.12.2 broke this
functionality. The folder 'exploded-aar' used to be in the
top-level 'build' directory and now it is in the module's
'build' directory, where it was excluded from the project.

Fixes issue
https://code.google.com/p/android/issues/detail?id=74341

UI test in
https://android-review.googlesource.com/103192

Change-Id: I17437180396406215f9dcb17430e8ee3ca095ce6
(cherry picked from commit 2e4028deab6d5e061fb364c2fad066add43dccfb)