navigator: Rework how the source provider is obtained for a resource
Res files are annotated with the source provider's name in the Android
project view. The provider corresponding to a res file was obtained
by looking at user data stored in the res file's containing res folder.

However, this mechanism of relying on user data on PsiElements may not
be a good approach since PsiElements may end up being recreated.

So rather than relying on this user data, we simply query the list of
source providers to identify which source provider contributed the
given res folder.

Change-Id: I83fba8affb8b1b45d590c0dc1aa5346a3aea2832