Toggling a form factor's selection state now triggers necessary logic.
On the "Select form factors" page in the "Create new project..." wizard,
changing the value of a form factor's API level pulldown would set the
page to a modified state, which, in turn, would cause some logic to get
run, doing such things as calculating the highest required API and
figuring out if we should show the license SDK page.

However, checking / unchecking form factors didn't set the page as
modified, so that logic was getting skipped.

This is now fixed.

Bug: http://b.android.com/73241

Change-Id: I7eccf44bea7ad3c87732b97d459fff47677d23ba