Bug 74220: Refactor to remove one thread.
DDM initialization involves following threads:
1. ADM monitor thread that monitors ADB process thread.
2. ADM monitor monitor thread - was spinning while ADB monitor
   thread was spinning.
3. ADM monitor monitor monitor task - cancelable task that runs
   in progress manager and reports indefinite progress while (2) is
   starting.
4. UI thread that waits for (3) to complete.

This change removes (2), making (3) track (1) startup instead. This
reduces the delay between the ADB startup and UI refresh. E.g. on my
system a modal progress indicator is not shown when opening
the Logcat view.

Note that this CL is a standalone refactoring and does not alter
the behavior. More CLs will follow.

Change-Id: I26d0d32e2cdfb173ecbebec843d7b3ef2c99eb67