Fixed test failures in AndroidProjectWizardTest
The tests in AndroidProjectWizardTest depended on some code in JetBrain's
version of the Android plugin that we didn't have, possibly due to a bad
merge earlier. I copied over the minimal changes required to get the tests
working again. In this case, JetBrain's version of a class that constructs
an Android module handled the creation of empty modules slightly differently
than ours did.

Change-Id: Ie479a4a24b314bf88e3731ebaab47e0bfbae5127