Fixed bug 73640 in JPS builder.
The bug is about the IDE invoking "assemble" Gradle
task for modules that are of "local AAR" type. This
type of module do not have the "assemble" task and
when building the project the IDE should just skip
them.

This bug was already fixed in the GradleInvoker
strategy (in-process build) in
https://android-review.googlesource.com/103541
but it was not fixed in the JPS build (out-of-process
build.)

The solution to the problem is to carefully
add the "Java" facet only to the non-Android modules
that have the "compileJava" task. Then we would only
build the modules with either the Android or Java
facet. The main problem with this solution is that
we need a "Java facet" counterpart in the JPS world.
When dealing with JPS each facet needs to have 2
copies: one for the IDE and one for JPS.

The need for the facets is to be able to figure out
which Gradle tasks to call. We were doing this on
both GradleInvoker and JPS, duplicating the logic on
both sides.

This CL introduces a better solution: all the Gradle
tasks to invoke are figured out in one place: the IDE.
Then the IDE will pass the list of tasks to JPS via
Java properties. This way, we have only one place
where this logic lives, and we can remove a lot of
duplicated code in the JPS side.

Finally fixes issue
https://code.google.com/p/android/issues/detail?id=73640

Change-Id: Ib22de9b654ca975f9ce66bcf80899ce2936b9a37