Now we remove all IDEA files before importing a project.
The motivation for this is to start a project import with
completely clean project state. When working on
https://android-review.googlesource.com/103541 I noticed that
certain things (e.g. facets) get added to an iml file but
not removed. For example, we stopped adding the "Java" facet
to "local AAR" modules. After fixing this, the test still
failed, because the .iml file still had the "Java" facet.

To delete the .iml files in a project, before importing it
(very important,) we parse the .idea/modules.xml file and obtain
the list of .iml files from there. Then we delete the .iml
files and the .idea folder.

Change-Id: I5504ed0f4e1b33542b2f0b486bce2479fbe85431