Fixed bug where AssertionError was thrown for project with custom location of build folder.
The problem is that when locating the content root parent of the build folder,
the check assumed that the folder was already created, which is not always the
case. For example, when importing a project that has not been built, the
build folder does not exist.

In this CL we check that if the build folder does not exist, we use the URL
of the folder specified in the content root to derive the path of the build
folder, this way we can properly set up the project "excluded" folders
during project sync.

This CL also includes a UI test and a project that reproduces this issue:
CentralBuildDirectoryTest and CentralBuildDirectory, respectively.

Fixes issue https://code.google.com/p/android/issues/detail?id=74259

Change-Id: If3eedf43c6eb86581dec815e16486211ea10f196