navigator: Customize representation for non Java modules
The Android view for non-Android modules simply defaulted
to the package view. The idea was that the package view was
written with a similar intention: to place more importance on
the logical view of the project and less importance to the
on-disk structure.

However, it doesn't work really well. Users find it confusing
that tests, sources and resources are all mixed together.

This CL provides some structure to such projects. The info
available for such projects are really just the kind of
source roots that are specified for the module, so the Android
view groups sources by the source root type (java, tests,
resources or test resources).

Fixes https://code.google.com/p/android/issues/detail?id=74732
and possibly https://code.google.com/p/android/issues/detail?id=74057

Change-Id: I380cc2db81521ff8ea71fb2eb5ea4955154dc394