Now we detect and fix if the wrong version of Gradle is used in the wrapper.
The next version of the Android Gradle plug-in (0.13.0) will require
Gradle 2.0 as a minimum. In this CL we detect if a project is using
plug-in 0.13.0 and check that the version specified in the Gradle wrapper
is 2.0. If the wrapper has an older version, the IDE will notify the
user about the new Gradle requirement and it will automatically fix the
wrapper before syncing the project with Gradle.

This CL only takes care of projects using the Gradle wrapper. To
properly ensure that the migration to 0.13.0 is smooth, the following
work needs to implemented in future CLs:
- Detect and fix Gradle version when using a local Gradle distribution
- Have a "quick fix" hyperlink in the "Messages" window, in case the
  IDE was not able to update the wrapper and sync failed due to a wrong
  version of Gradle
- Change the "recommended" version of the plug-in to be 0.13.0
- Update the current "quick fix" to update both the version of the
  plug-in and Gradle.

The work in this CL required to reuse ProjectValidator. As part of this
CL I also fixed this related, existing bug
https://code.google.com/p/android/issues/detail?id=68698

Change-Id: Ia0b825383ce45753380a06dc6ad6bfc684ae48d3