Added support for checking minimum Gradle version to ChooseGradleHomeDialog.
This CL adds checks for a minimum Gradle version to the "Select Gradle Home"
dialog. After a user selects a valid home directory, the dialog will check
that the selected directory contains a Gradle distribution with version equal
or greater than the specified minimum one.

This work is necessary to support the migration to Android Gradle plug-in
0.13, which introduces a dependency on Gradle 2.0 or greater.

This CL contains both unit tests and UI tests.

Change-Id: I82213277ee1dfbaf7c14c7eeab2a933bcbe879b5