[strings-editor] Removes buggy save/restore selected cell functionality.
Removes the code to remember the currently selected cell in the table
and reselect it after updating the string resource data.  The
implementation failed to reselect the cell in some corner cases (e.g.,
the first time the user makes an edit) and did not work reliably when
cell editing was enabled.

Change-Id: Ia7ed6304b45f75780f7536231c421306567f7008
Signed-off-by: Rebecca Chen <rechen@google.com>