[strings-editor] Makes table cells directly editable.
Turns on editing for all table cells and replaces the default cell editor
with a JTextArea inside a JBScrollPane to hold values that can span
multiple lines.

Known issues: If you click on a cell that is not wide enough to show all
of its text, a single click starts editing. But if you click on a cell
wide enough to accommodate all of its text, a double click is needed.

Change-Id: I204964dd3fdaddcbf7e85120ff1212f49298ad2d
Signed-off-by: Rebecca Chen <rechen@google.com>