Fixes to Navigation Editor for ioshed app - continued
Better default layout.

The problem with the existing layout strategy is that apps like ioshed
have lots of states with no transitions associated with them. These get
racked up vertically and the overall layout is then too high to effectively
scroll through.

This CL addresses this as follows:

Fish out all the states (Activities/Menus) that are not involved in
any transitions and add them to the bottom of the layout. These
'island' states are then presented in overlapping form so they take
a small amount of real estate and the overall layout is a 
manageable starting point for user edits.

Change-Id: I07cbf327e6a84d81229afa89cb80ed156544c1a6