Added infrastructure to UI testing to simulate Gradle sync failures.
There are bugs that happen during Gradle sync that are very
difficult to reproduce. One example is
https://code.google.com/p/android/issues/detail?id=74842
where apparently Gradle did not finish downloading artifacts
and now project sync fails.

In this CL adds a "execute before sync" task which can be
used to simulate failures in Gradle. See
CorruptedGradleDependencyTest for more details.

This CL also introduces the concept of "isGuiTestingMode", similar
to IDEA's "isUnitTestingMode".

Change-Id: I27b42a68cbb012696b93b2745092e39458a7d15c