Now we detect if the Gradle cache is corrupt or out of sync and attempt to fix it.
If for some reason Gradle failed to download dependencies correctly, project
sync will fail with the message "Premature end of Content-Length delimited message
body".

In this CL we detect that, show an user-friendly message and include a hyperlink
that offers the user to sync the project passing the "--refresh-dependencies"
command line option.

Fixes issue
https://code.google.com/p/android/issues/detail?id=74842
and sets up infrastructure for fixing
https://code.google.com/p/android/issues/detail?id=73872

Change-Id: I5e8b61bba7bae64bf5a89135d74fd8c0e72bc1ec