strings editor: use font that can best display each string
On Windows and Linux, the default code editor font does not
support various languages. The code editor however, works around
this by using the default font wherever possible, but for
characters that aren't handled by the default font, it picks
a fallback font that can handle that character.

This CL implements a similar methodology for the cell renderer
in the Strings editor.

Fix for https://code.google.com/p/android/issues/detail?id=75196

Change-Id: Ibada09f95929918e1420150ea94ae3b56a7d2164