Change memory profile window to use user triggered start and stop.
Adds a recording toggle action to the toolbar that starts/stop a profiling.
Now when the tool window is minimized the profiling continues in the
background. In order to fully terminate the profiling session, there is a new
action (cross) to do so. This is more inline with other tool windows in
intellij.

MemoryProfilingView is now in charge of choosing the Client to sample and
MemorySampler listens for changes done to a DeviceContext.

Change-Id: I41b28df80e200a8024fd840790da9bdb756660fa