Decouple wizard from DialogWrapper.
"We can solve any problem by introducing an extra level of indirection."

For our new welcome wizard, I want to reuse existing wizard code
from the "welcome frame" (it is a window that appears when no
projects are open).
This change decouples the dynamic wizard from the DialogWrapper that
is used to create IDE dialogs. In another branch, I have WIP
implementation of he wizard embedded in the welcome screen. It should
also be possible to embed wizards into IDE panes or editors.

Change-Id: I087116626cfd707c3111676725bdf4d51a9f5d7d