Remove unnecessary preference pages.
With CLs
https://android-review.googlesource.com/94253 and
https://android-review.googlesource.com/94280
we were already removing "Preferences" that did not
belong to Gradle-based Android projects (e.g. JavaFX,
Gant, etc.)

Even though the code that removed this options is
correct, "when" this code was invoked was not right. These
preference pages were removed *after* a project sync
finished (successfully or not.)

This CL fixes this by removing the unnecessary
preference pages:
1. When sync starts
2. When project state is loaded from a disk cache

In addition, I renamed the "Compiler" page for Gradle-based
Android projects to "Compiler (Gradle-based Android
projects.) The reason is for non-Gradle projects and
the "Welcome" page there are 2 "Compiler" entries (the
default IDEA one and the one we add,) which can be confusing
to our users.

Fixes issues
https://code.google.com/p/android/issues/detail?id=75424
https://code.google.com/p/android/issues/detail?id=75426

Change-Id: Id2430be5f47c448965ff502f03a143b01ae93594