Fix SDK_INT conditional checks for the lint API check
Lint will check to see if a call reference is inside a "if (SDK_INT >= n)" check,
which would make the API call safe if the call API version <= n. This logic
was accidentally reversed, but all the unit tests only checked the case where
the call version was the same as the version in the check, so the comparison
direction didn't matter. This fixes the checks and adds more tests.

It also makes the code which checks a class reference also look in the surrounding
context for an if-conditional, and it fixes the Add TargetApi quickfix such that
it works correctly when passing through anonymous classes.

Change-Id: I858604823d085d90b52bcd037e2cb889570e08c4