strings editor: save values from text fields on every keypress
The strings editor provides two ways to change the data in a
particular cell: Either by double clicking the cell, or by editing
the value in the text box at the bottom.

When the value is edited in the text box, the model is updated when
the text box loses focus. This CL changes it so that the model is
updated on every key press, so the table is updated as soon as the
value in the text box is changed.

Change-Id: Ib70502474bbf1802443819d5233f79fdfa42688e