Add internal actions to start/stop adb + minor refactoring
This CL adds a couple of internal actions (that show up only in debug
builds under Tools | Internal Actions | Android) to start/stop adb.

The method to get the debug bridge now takes a File parameter instead
of a String, and AndroidSdkUtils exposes a way to get the file corresponding
to adb.

Change-Id: Ie715822842b77a7ed97ad4934439d09491173480