Add a spell checking strategy for text and property files
In a default project, if you run Analyze > Inspect Code..., you end up
with about 80 spelling mistakes from non-user-code files like the
gradlew and gradlew.bat files (listing lots of BAT file keywords,
partial command line flags, etc).

This CL adds a new spell checking strategy for text and property files
which is inserted into the front of the spell checking strategy lists
for these file types, and this special strategy implementation decides
whether to mask errors from a given file based on the filename, and if
it's an ignored file, it returns a no-op tokenizer.

This removes all the false positives from the gradle wrapper files
(and .gitignore, and gradle.properties, etc.) It also adds a couple of
words to the default dictionary which occur in default projects (e.g.
"dimens" is mentioned in the default dimens.xml file as a comment, and
AppCompat is used in the default style which is tokenized to the words
and app and compat.)

Change-Id: Ia586aa4f5c80540ff2094fc83ba0829aca2e5ac0