navigator: Open appropriate manifest on double click
When the manifests group node is double clicked, currently the tree
expands to reveal the list of manifest files. This CL changes the behavior
to open the default manifest instead. This is similar to 7da038b
that changed the behavior of double clicking resource group nodes to
open one of the resource items instead.

Fixes https://code.google.com/p/android/issues/detail?id=75912

Change-Id: I6c78c1c0e1c46a281bfddf99437148fb8ac7cf84