73991: Adds ApkInstallerService to not upload an apk twice
AndroidStudio will now push only the APKs that have changed.
This is done by keeping a cache in memory of the APKs hashcode that
where pushed to a specific location on the device.

If the APK content changed or the device is disconnected the cache
is cleared.

This is silimar to what we did in Eclipse.

See https://code.google.com/p/android/issues/detail?id=73991

Change-Id: I6d7f67b20819c6ea864361e068204f4e8063da78