strings editor: Add support for editing multiline strings
This CL enables the multi line editor to show up when pressing
Shift+Enter in the details text box (or pressing the corresponding icon).

It also disables editing of multi line strings in the text box itself,
and also disables editing of keys. Editing of keys should actually be
a rename refactoring and will be enabled in a future CL.

Change-Id: Ib9532d9f5e66ee1e4d6acdbfbea58f790ab5936a