Add rendering tests
This CL adds infrastructure for easily testing the visual appearance of layout
renderings. You can require that the rendering should be roughly similar to
an existing thumbnail of the same scene (and the first time you run the test
it will generate this thumbnail such that you can visually inspect it and make
sure it looks correct). On subsequent runs the rendering is painted, scaled
and compared to the golden file and the test passes if the image is sufficiently
similar to the rendering.

If the images are too different, the test will generate a "delta" image which
shows the before, after and difference.

The test currently checks a couple of key renderings:
- a basic layout with an action bar
- a layout including most of the widgets
- a layout with some gradients and included layouts
- a layout rendered as included in another (e.g. testing overlay masks)
- a drawable resource
- a menu resource

Change-Id: Ic6f32a00346ab89a0569e49ceaac89d257ab69e9