Added hyperlink to migrate to Gradle wrapper if sync fails because of unsupported version in local Gradle distribution.
This is part of the work related to migration to Android Gradle plug-in 0.13.

If Gradle sync fails because project is using plug-in 0.13 with an unsupported
Gradle version (1.*) and the user did not set the IDE to use the right Gradle
version (the IDE prompts user to do this automatically before sync, but user
can opt out,) the error message will show a hyperlink that migrates the project
to the Gradle wrapper using the right version of Gradle, and automatically
syncs project after that.

Change-Id: I4d42829bfb83a5aac19810a96cd7de9bbc3c2c25