75983: Sync notification doesn't trigger for all edits
There were two bugs that needed to be fixed:

1) We need to not just check file stamps for build.gradle files, but
   also the document registry to see if the .gradle file has been
   edited but not yet saved

2) We need to trigger the editor notification check not just for the
   PSI childReplaced event, but for childAdded and childRemoved as
   well.

This CL also adds GUI testing support for the editor notification
facility; it can now check that there are no editor notifications, or
check that a particular notification is showing, and finally it can
click specific hyperlinks in the notifications. The regression test
for this bug performs all three.

Change-Id: Id07f10e5cc412a518da2ea56e6c4196c6f86b97a