[UI Testing] Fixed GradleSyncTest.
Execution of tests in GradleSyncTest was getting stuck when
running testUnsupportedGradleVersion. The reason is that the
IDE's file system needed to be refreshed before finding the
VirtualFile that corresponds to the file to select in the
"Open Project" dialog, and refreshing the local file system
while the modal dialog is up resulting in the UI to freeze.

The fix is to refresh the local file system (if necessary)
before invoking the "Open Project" dialog.

Change-Id: I0cc893d5375f549aee4f478334b07cf3872fbe97