[UI Testing] Now we use the "Open Project" wizard only once.
With this CL, we only use the "Open Project" wizard the first
time we open a project in the test suite. After that, we open
projects through APIs.

This will make test runs a little faster. It may also fix
an issue where opening a project through the UI freezes a test
when trying to refresh the local file system.

Change-Id: Id6d01bcb6ba07ef90f1c4dadc0677a95b3c55e9f