navigator: Show project as folder if gradle sync has failed
The Android Project view will not show any modules that don't have
a single source root. This is usually not a problem, except if there
are errors during gradle sync, we explicitly create a new project
without any source roots, in order to allow users to edit that project.

This CL adds support for such a scenario: If we notice that the project
is built by Gradle, and the sync failed, then we add the top level
project directory.

Fixes https://code.google.com/p/android/issues/detail?id=76564

Change-Id: I43ab46531c5f22db640a103b63edebf022c4f8bd