76715: Generated resource files are flagged with warnings
This CL hides the warnings described in issue 76715 (well, the spell checker
warnings; the unbound namespace is hidden by a separate fix for
XmlUnboundNsPrefixInspection). It also adds GUI test infrastructure
to run inspections on projects and to assert a given expected result.
This is used to verify that a newly created project, with resValues
added to the build.gradle, has no unexpected warnings.

Change-Id: I227947cdb2e631c6cfb8e88c5ad3a6539145fc9e