Add extension point for allowing tests to be run on the cloud
This CL adds a CloudTestConfigurationProvider extension point. This allows
users to run tests on the cloud on a set of specified device configurations.
At a high level, a cloud provider is expected to provide a way to select
a cloud project, and a set of device configurations on which the tests
ought to be run.

The test run config editor exposes these options to the user. Selecting
cloud as the target in the run config will result in tests being run on
the specified cloud provider (the actual execution is delegated to the
provider).

This feature is disabled by default. It requires a plugin that supplies
the extensions, and setting -Dshow.google.cloud.testing.option=true

Change-Id: I0402152fd3b26efc8f0f7f6dde110d10c41acaea